Policy Refinement Using Bayesian Optimization/requirements.txt
python==3.8.5 
gym==0.18.0 
numpy==1.20.3 
pytorch==1.8.1 
Gpy==1.9.9 
GPyOpt==1.2.6
pyyaml
